+++ src/de.hpi.swa.graal.squeak/src/de/hpi/swa/graal/squeak/nodes/context/stack/StackPeekNode.java
@@ -22,7 +22,7 @@ public abstract class StackPeekNode extends AbstractStackNode {
     @Specialization(guards = {"isVirtualized(frame)"})
     protected Object doPeekVirtualized(final VirtualFrame frame) {
         CompilerDirectives.ensureVirtualizedHere(frame);
-        return readNode.execute(frame, (int) frameStackPointer(frame) - offset);
+        return readNode.execute(frame, frameStackPointer(frame) - offset);
     }
 
     @Specialization(guards = {"!isVirtualized(frame)"})
